Adjust hero image
...well this was a learning experience...will
never use periods in a class or i think at
this point ANY attribute. Spent way too long
trying to fix the css and thinking i didn't
get the concept as well as i thought before
realizing it was an issue with the selector.
which i did check before but did not think
the issue was this..anyways now to get the
side done woot!